Bugfix `Nudge` duplication
Prior to this fix, `Entry`ies flagged as `Nudge` by the `Deduplicator`
would still be delivered to the application layer. This is because only
`Entry`ies flagged as `Ignore` would be removed from the burst after the
amended root is complete (note that `Drop` `Entry`ies would be removed
from the burst before the amended root is computed). This fix promotes
`to_ignore` to `to_omit`, and fills `to_omit` with `Nudge` indices too.